Simplify HTTP: Separate bastion registration
